front/requirements.txt
absl-py==0.5.0
agate==1.6.1
agate-dbf==0.2.0
agate-excel==0.2.2
agate-sql==0.5.3
appnope==0.1.0
astor==0.7.1
Babel==2.6.0
backcall==0.1.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.3
bleach==2.1.4
boto==2.49.0
boto3==1.9.58
botocore==1.12.58
bpemb==0.2.11
bs4==0.0.1
bz2file==0.98
CacheControl==0.12.5
cachetools==3.1.0
certifi==2018.8.24
chardet==3.0.4
cheroot==6.5.2
CherryPy==18.0.1
Click==7.0
csvkit==1.0.3
cycler==0.10.0
dbfread==2.0.7
decorator==4.3.0
defusedxml==0.5.0
Deprecated==1.2.5
Django==2.1.7
django-widget-tweaks==1.4.1
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fasttext==0.8.22
firebase==3.0.0
firebase-admin==2.16.0
flair==0.4.1
Flask==1.0.2
future==0.16.0
gast==0.2.0
gastrodon==0.9.3
gcloud==0.17.0
gensim==3.7.1
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
grpcio==1.12.1
h5py==2.8.0
html5lib==1.0.1
httplib2==0.12.1
hyperopt==0.1.2
idna==2.7
image==1.5.27
ipykernel==5.0.0
ipython==7.0.1
ipython-autotime==0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isodate==0.6.0
itsdangerous==1.1.0
jaraco.functools==1.20
JCC==3.3
jdcal==1.4
jedi==0.13.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.3.0.dev0
jupyter-core==4.4.0
jws==0.1.3
Keras==2.2.3
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
leather==0.3.3
lucene==7.5.0
lupyne==2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.3
mkl-fft==1.0.6
mkl-random==1.0.1
more-itertools==4.3.0
mpld3==0.3
msgpack==0.6.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
nltk==3.3
notebook==5.7.0
numpy==1.15.2
oauth2client==3.0.0
openpyxl==2.5.9
pandas==0.23.4
pandocfilters==1.4.2
parsedatetime==2.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
portend==2.3
prometheus-client==0.3.1
prompt-toolkit==2.0.5
protobuf==3.6.0
ptyprocess==0.6.0
py4j==0.10.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pybind11==2.2.4
pycryptodome==3.4.3
Pygments==2.2.0
pymongo==3.7.2
pyparsing==2.2.2
Pyrebase==3.0.27
pyspark==2.3.2
python-dateutil==2.7.3
python-jwt==2.0.1
python-slugify==1.2.6
pytimeparse==1.1.8
pytorch-pretrained-bert==0.6.1
pytz==2018.5
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
rdflib==4.2.2
regex==2018.1.10
requests==2.21.0
requests-toolbelt==0.7.0
rsa==4.0
s3transfer==0.1.13
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
segtok==1.5.7
Send2Trash==1.5.0
sentencepiece==0.1.8
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
smart-open==1.7.1
SPARQLWrapper==1.8.2
SQLAlchemy==1.2.12
sqlitedict==1.6.0
tempora==1.13
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
torch==1.0.1.post2
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
Unidecode==1.0.22
uritools==2.2.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
xlrd==1.1.0
zc.lockfile==1.3.0
